fix(coredump): serialize stack values as i32 to match declared tag
Corestack entries were still writing 8 bytes per value while tagging
them 0x7F (i32), leaving every downstream entry misaligned by 4 bytes.
Truncate the copy to sizeof(uint32_t) so the buffer size, memcpy
width, and the i32 tag all agree.

+++ lib/executor/coredump.cpp
@@ -129,8 +129,8 @@ AST::CustomSection createCorestack(
       // support i64.
       Content.push_back(0x7F);
       auto Value = Iter.unwrap();
-      std::vector<Byte> ValueBytes(sizeof(int64_t));
-      std::memcpy(ValueBytes.data(), &Value, sizeof(int64_t));
+      std::vector<Byte> ValueBytes(sizeof(uint32_t));
+      std::memcpy(ValueBytes.data(), &Value, sizeof(uint32_t));
       Content.insert(Content.end(), ValueBytes.begin(), ValueBytes.end());
     }
     if (!ForWasmgdb) {
@@ -139,8 +139,8 @@ AST::CustomSection createCorestack(
         // support i64.
         Content.push_back(0x7F);
         auto Value = Iter.unwrap();
-        std::vector<Byte> ValueBytes(sizeof(int64_t));
-        std::memcpy(ValueBytes.data(), &Value, sizeof(int64_t));
+        std::vector<Byte> ValueBytes(sizeof(uint32_t));
+        std::memcpy(ValueBytes.data(), &Value, sizeof(uint32_t));
         Content.insert(Content.end(), ValueBytes.begin(), ValueBytes.end());
       }
     }
